Increase the wait period after starting Minio. (#24693)
In one of the sanities, minio panicked becuase globalLifecycleSys was
not initialized. According to the code and unit testing, it gets initialized
as part of the main function, since minio.Main is called in a separate go routine
its possible that globalLifecycleSys was not initialized by the time SetBucketLifecycle
was called.