Increase retryIntvl for indexing fwlogs. (#24767)
Change log level from Info to Debug. It was changed to Info for debugging another issue sometime back, moving it back to Debug otherwise it fills up logs. After the final failure (if any) another err
log will get printed.

If the retry interval is too small (current value is 100ms) then it compounds the load on Elastic since the whole batch gets retried. Changed it to 500ms for fwlogs.