ionic.spec: save inbox drivers and update initramfs on rpm install (#25326)
If the RPM is installed on a system that already has an ionic
driver, save the original driver so it can be restored if the
RPM is later removed.

When installed or removed, update the initramfs so that the
driver is available in the early boot time.

When installing and removing the RPM, don't have the RPM do the
driver load or unload.  We need to let the customer decide when
is the best time to interrupt driver operations.  This allows them
to preload many bits of SW and then do a single scheduled down time
as needed.

Other little spec cleanups while we're here.

Signed-off-by: Shannon Nelson <snelson@pensando.io>