Fix for TCP traffic loss during vMotion (#25896)
During ENIC create, entry in the Input Properties MAC VLAN with skip_flow_update will be added.
At start of vMotion (in EP Create), this entry will be deleted (and later added in the vMotiion end)
During vMotion - there is a small window where the flows will be skipped from flow update (SeqNum/AckNum
will not be updated). Later when that entry is deleted, because of SeqNum/AckNum is out of the window,
packets could be dropped.

Fix is to - During ENIC Creation time itself, set a flag when the entry is created as part of vMotion.
Then dont programm the Uplink entry itself. Later when the vMotion completes, add the entry.