Fix ListObjects bug in Minio. (#26166)
When a node reload happens, minio still keeps accepting new data. The new data is not immediately synced with the node that was reloading, instead a heal operation is run once in 30 days for syncing
the data. The query operation continue to work becuase the data is still present on other nodes.

As part of the PR https://github.com/minio/minio/pull/9184 minio started using 3 random erasure coded disks for listing objects and started relying on disk healing for consistent results.
This approach does not work correctly becuase heal runs once in 30 days and if node reload happens in between then listObjects API will start giving inconsistent results depending on which random disks
are picked for the list operation.

As part of this patch we are disabling the usage of 3 random disks for list operation.

Did some refactoring in test cases as well.