Fix TestEventsAlertEngineWithAPIServerShutdown (#26191)
Added the following fixes for TestEventsAlertEngineWithAPIServerShutdown
failures seen in recent jobd runs.

1. jobd #5435684: Expected alerts are not created for the alert policy used in the test.

This occurs because the API server is stopped before the alert policy,
added by the test, is seen by evtsmgr and added to its memdb cache. A
1-second delay is now added between adding the policy and stopping the
API server.

2. jobd #5022454, #5449935: Some expected alerts are not seen.

From the logs, 4 alerts are created - 2 each for the default
event-based-alerts policy and the policy added by the test. But the test
times-out before they are all pushed to the API server.

The wait time in the test is now increased to 20 secs.

3. jobd #5239634: Failed to start API client.

Added AssertEventually to make multiple attempts to start the API client
in the test.

4. jobd #5339591: AutoAddAlert(), AutoGetAlert(), AutoUpdateAlertPolicy() fail with deadlineExceeded RPC error

From the logs, these functions take more than 2 seconds to complete. So,
increased the timeout values for these functions in createAlert() and
updateAlertsHelper().

Co-authored-by: Sudheendra Gopinath <sgopinath@pensando.io>