linux ionic: no memory realloc on link changes (#25910)
We've found that there are memory fragmentation issues inside
Naples, and the driver's propensity for freeing and re-allocating
all the queue structures on link flaps can aggravate it to the
point of failure.  This reflects the feedback we got with this
from the upstream community when we first pushed the fw_upgrade
changes.  The result in the upstream driver was a rewrite of
some of the ionic_open() and ionic_close() activities to create
a new layer that uses ionic_start_queues() and ionic_stop_queues()
across a link flap in order to not need to free and re-alloc all
the memory.

This patch converts our out-of-tree driver to use that same
method.

Signed-off-by: Shannon Nelson <snelson@pensando.io>