pcieport: suppress rxtlp_err log events (#26255)
Sometimes the host gets into a state where it streams some tlps
to us that we don't handle.  These generate an rxtlp_err_interrupt
event which we count in portstats.  No need to generate a log event
for this interrupt.

Also protect pcieutil healthdetails when run with no pciemgrd started.