ionic.spec: untangle the rpm-U intall-uninstall clash (#26402)
The "rpm -U" first installs the new, then removes the old.  If the file
names are the same, the 2nd step can undo the 1st step's actions and end
up with the new rpm using the old ionic.ko.  We need to add some version
checking to prevent the tangled install.

Instead of installing "ionic.ko", we install "ionic.ko-{version}" which
will be tracked by the RPM system.  Then we link that to ionic.ko for
general use by the kernel.  When uninstalling, the RPM system will
remove "ionic.ko-{version}", and we need to clean up the ionic.ko file,
but only if it is still linked to "ionic.ko-{version}".  In this way,
the rpm -U command can install the new version and the remove of the
old version won't disturb the ionic.ko.

We also now only install the udev hack for the RHEL versions needed.

double commit #26340

Fixes: PS-3531, PS-3700

Signed-off-by: Shannon Nelson <snelson@pensando.io>