linux ionic: double commit collection (#26679)
* reduce max Rx ring size and add error recovery (#26259)

* linux ionic: recover from ringsize change failure

If the ringsize change fails, try restoring the previous setting.
This also fixes up a couple issues in the error recovery path.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: limit rx ringsize to 8k

The recent performance patches increased the size of struct ionic_desc_info
such that the alloc request for the q.info for 16k ring size is greater
than 4Mbytes and the allocation fails.  We can limit the ring size to
8k to prevent this.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: keep rss hash after fw update (#26420)

Make sure the RSS hash key is kept across a fw update by not
de-initing it when an update is happening.

Fixes: PS-3751
Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: clean up the dbid_inuse_lock on lif_deinit

The mutex_destroy() was missing for dbid_inuse_lock.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: use mutex to protect queue operations

The ionic_wait_on_bit_lock() was a open-coded mutex knock-off
used only for protecting the queue reset operations, and there
was no reason not to use the real thing.  We can use the lock
more correctly in the Link status check to better protect
the queue stop and start operations from cross threading.
We can also remove a useless and expensive bit operation from
the Rx path.

Signed-off-by: Shannon Nelson <snelson@pensando.io>