Vs2062prime (#26531) (#26745)
* Fix regression in disk monitor logic.

Add options for testing and fix regression in disk monitor logic.

* Fix for VS 2062

The file name change done for including vrfName  exposed an issue in fwlogs processing.
Spyglass depends on lastProcessedKey file to figure out the last object processed for a DSC.
If Spyglass restarts, it uses that file to figure out the starting point. In this issue,
since the file name changed to include vrfName in it, it resulted into a state where it started
listing from an older date and got stuck. The system is in a state from where it can't be
recovered because even if I delete the lastProcessedKey file, it will start listing from the
beginning (which may never end).
The fix involves the following changes:
1. If the file is not present then just process the new logs (recovery action)
2. If any error is seen while parsing the file then just process the new logs (recovery action)
3. Update the lastProcessedKey file even if logs are getting rate limited because from Spyglass's
perspective, it processed that file once and rate limited it.
4. Don't relist objects of a dsc if the last processed object is older then 1 hour.
5. Use a worker pool for listing objects in Vos. Before this fix, it was spawning a new routine
for every DSC.

Additional change:
1. Index starts and endts properties in fwlogobjects bucket.
2. Ignore deleted fwlog object notifications. Delete notifications are generated when objects are deleted by the disk cleanup routine.