rpms: add sles flag to force new initrd (#27855)
The kernel_module_package in sles rpm needs a '-b' argument in order to
force the dracut run to get a new initrd, when the module being added
is not already known to the initrd.  This '-b' argument is not supported
by the rhel version of the macro, so can only be used when building the
sles rpms.

Since the macro also replaces the post scriptlet, we have to do the
udev rule hack a little differently.  Here we make it a full fledged
file for the 3 rhel versions that need it.

While we're here, we can remove the other scriptlet sections that are
overridden when using the kernel_module_package macro.

Signed-off-by: Shannon Nelson <snelson@pensando.io>