linux ionic: No full driver reset on config changes (#28144)
* linux ionic: clean up unnecessary non-static functions

ionic_open() and ionic_stop() are not referenced outside of their
defining file, so make them static.

Also drop a left over field from the qcq struct.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: use kcalloc for new arrays

The upstream folks have reminded us that using kcalloc is proper form
for allocating arrays of structures.

We can fix up a couple of debug strings while we're here.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: change mtu without full queue rebuild

We really don't need to tear down and rebuild the whole queue structure
when changing the MTU; we can simply stop the queues, clean and refill,
then restart the queues.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: change the descriptor ring length without full reset

The original way of changing ring length was to completely tear down
the lif's queue structure and then rebuild it, while running the risk
of allocations that might fail in the middle and leave us with a broken
driver.  This in fact bit us in some Naples stress test situations.

Instead, we can set up all the new queue and descriptor allocations first,
then swap them out and delete the old allocations.  If the new allocations
fail, we report the error, stay with the old setup and continue running.
This gives us a safer path, and a smaller window of time where we're not
processing traffic.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: change queue count with no reset

Add to our new ionic_reconfigure_queues() to also be able to change
the number of queues in use, and to change the queue interrupt layout
between split and combined.

We drop the queue-split private flag from ethtool in favor of the more
correct way by using ethtool -L's rx and tx values versus the combined
value for spitting or sharing interrupts between the queues.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: pull reset_queues into tx_timeout handler

Convert tx_timeout handler to not do the full reset.  As this was
the last user of ionic_reset_queues(), we can drop it.

Signed-off-by: Shannon Nelson <snelson@pensando.io>

* linux ionic: more macro cleanup

There are a few more bits to clean up after changing the memory
layout, mainly removing some unnecessary macros.

Signed-off-by: Shannon Nelson <snelson@pensando.io>