Make fast flow logs index vpc aware. (#28317)
The index has an extra hierarchy for vpcs.
There is no change in the way index files etc. are managed. The indices for the all the VPCs under a tenant are kept together
in the same set of files. This is different then how raw logs are stored, where the vpc name is present in the file name itself.